+++ lib/idkit_gshared.dart
@@ -1,1 +1,3 @@
 library idkit_gshared;
+
+export 'src/gshared_main.dart';